+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Mar 7
 * Added a progress bar for better UX
 * Supress the annoying `find: ‘.../lost+found’: Permission denied` - which is a system directory anyway
+* Added a `Subtitles` column
 # Tag v3.1.0
 ## Mar 5
 * Migrated from `ffrpobe` to `mediainfo` (dependenciesa changed)